feat: entities/models/*.types 분리
