+++ src/entities/comment/models/comment.types.ts
@@ -1,0 +1,10 @@
+type Comment = {
+  id: number;
+  postId: number | null;
+  body: string;
+  likes: number;
+  userId: number;
+};
+type NewComment = Omit<Comment, "id" | "user" | "likes">;
+
+export type { Comment, NewComment };
