--- src/entities/comment/types/comment.ts
@@ -1,1 +1,0 @@
-//
