--- src/entities/post/types/post.ts
@@ -1,1 +1,0 @@
-//
