--- src/entities/user/types/uset.ts
@@ -1,1 +1,0 @@
-//
